+++ mvc/src/main/java/com/interface21/webmvc/servlet/adapter/ControllerInterfaceHandlerAdapter.java
@@ -1,23 +1,27 @@
-package com.interface21.webmvc.servlet.mvc.fake;
+package com.interface21.webmvc.servlet.adapter;
 
+import com.interface21.webmvc.servlet.HandlerAdapter;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-public class FakeHandlerAdapter implements HandlerAdapter {
+public class ControllerInterfaceHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(Object handler) {
-        return handler instanceof String;
+        return handler instanceof Controller;
     }
 
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler)
             throws Exception {
-        View view = new JspView("/jazz");
+        Controller controller = (Controller) handler;
+        String viewName = controller.execute(request, response);
+        View view = new JspView(viewName);
+
         return new ModelAndView(view);
     }
 }
