+++ mvc/src/test/java/com/interface21/webmvc/servlet/HandlerMappingRegistryTest.java
@@ -1,10 +1,10 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.interface21.webmvc.servlet.mvc.fake.FakeHandlerMapping;
+import com.interface21.webmvc.servlet.fake.FakeHandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
