+++ mvc/src/test/java/com/interface21/webmvc/servlet/fake/FakeHandlerAdapter.java
@@ -1,26 +1,23 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.fake;
 
+import com.interface21.webmvc.servlet.HandlerAdapter;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
-public class ControllerInterfaceHandlerAdapter implements HandlerAdapter {
+public class FakeHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(Object handler) {
-        return handler instanceof Controller;
+        return handler instanceof String;
     }
 
     @Override
     public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler)
             throws Exception {
-        Controller controller = (Controller) handler;
-        String viewName = controller.execute(request, response);
-        View view = new JspView(viewName);
-
+        View view = new JspView("/jazz");
         return new ModelAndView(view);
     }
 }
